Improve performance by reusing the gRPC connection
- Added pay.LNclient interface
- Added ln.LNDclient implementation of the LNclient interface
- Added a factory function for LNDclient
- Removed one and changed two functions in the ln package

The changing of the two functions in the ln package is what made the gRPC connection reuse possible.
And to pass a single reusable client object instead of multiple parameters (lnd client, context and gRPC connection would have been required) made much more sense,
leading to the addition of the interface and struct.

Closes #4