Add bbolt as local storage option
- Formerly known as Bolt / Bolt DB
- Added pay.BoltClient, which implements pay.StorageClient
- Added factory function NewBoltClient(...)
- Added pay.BoltOptions and pay.DefaultBoltOptions

This storage client implementation opens or creates a DB file and a so called bucket,
and then either concurrently reads key/value pairs or non-concurrently writes new key/value pairs.
The exclusive write is achieved with a sync.Mutex.
The DB is never closed, because it's used by the middleware until the web service stops
and it seems that as soon as the web service stops, the file lock is released automatically.

Closes #3