Fix a client can cheat with using the wrong preimage for a request
- Previously a client just had to send any preimage that corresponds to a settled invoice in the LN node,
without the invoice having to originate from the middleware, or from the same endpoint as used in
the current request. So he could pay cheap invoices and use expensive endpoints for example.
- The fix required storing metadata in the DB, e.g. the HTTP method and URL path,
while being backward compatible and not send those data encrypted as token in the first response,
requiring the same token and decrypting and using the values during the second request
- Both modes are interesting and have their pros and cons, so evaluate supporting both in the future
- Storing the metadata required changing the wall.StorageClient interface and its implementations
in the storage package

Closes #16